Make verify suite a function & pass parameters to it.
